export keys: working implementation.
Code is based on `signing.js` and is incomplete (cannot select which
keys to export).

Upon successful password entry, the keys are displayed in

<address>,<secret exponent>

form in the "tools" output box in the middle of the screen, one key per
line. Main hierarchy keys in the extension store are exported, including
"pocket" ones, but Stealth keys are not.