tools/exporting: disable export button after export is complete, move close-button code to a function, clean up HTML.
A non-disabled export button after a successful export suggests that
there is something left to be done, as in saving to disk. A better
approach would perhaps be hiding the button altogether, and showing
another one (e.g. copy to clipboard) instead.

If an invalid password is entered, the export button stays active,
allowing another attempt.